db first migration
add order base, change phone field type

+++ app/models.py
@@ -7,5 +7,5 @@ class Order(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     customer = db.Column(db.String, index=True)
     address = db.Column(db.String, index=True)
-    phone = db.Column(db.Integer, index=True)
+    phone = db.Column(db.BIGINT, index=True)
     eggs_count = db.Column(db.Integer, index=True)
